check if there any old kernels

+++ src/dnf.sh
@@ -5,7 +5,14 @@ echo -e "\n\n _____ DNF Cache _____"
 echo -e "\nSize of cache: $(sudo du -sh /var/cache/dnf)\n"
 
 # Remove old versions of kernel
-sudo dnf remove -y $(dnf repoquery --installonly --latest-limit=-2 -q)
+if [ -n "$(dnf repoquery --installonly --latest-limit=-2 -q)" ]; then
+  echo "Old kernels found:"
+  dnf repoquery --installonly --latest-limit=-2 -q
+  sudo dnf remove -y $(dnf repoquery --installonly --latest-limit=-2 -q)
+else
+  echo "No old kernels found."
+fi
+#sudo dnf remove -y $(dnf repoquery --installonly --latest-limit=-2 -q)
 
 # When upgrading to a new version of Fedora, a cache is created. In theory, the cache is cleaned after the upgrade. If not, cleaning can be forced using the following command:
 sudo dnf autoremove -y
